requirements.txt
python-can
socket
json
argparse
logger
packet
pyserial
requests